requirements.txt
## Hybrid RAG Analysis Requirements

# Retrieval
rank-bm25>=0.2.2
kiwipiepy>=0.17.0
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # GPU 사용 시 대체

# Embedding
sentence-transformers>=2.2.2
torch>=2.0.0

# Data Processing
langchain-text-splitters>=0.0.1
tqdm>=4.65.0
pyarrow>=15.0.0

# Utils
numpy>=1.24.0
pandas>=2.0.0

# Optional: LLM Evaluation
openai>=0.27.0  # OpenAI API
requests>=2.31.0
